+++ src/Phan/Plugin/Internal/MethodSearcherPluginLoader.php
@@ -1,0 +1,4 @@
+<?php declare(strict_types=1);
+namespace Phan\Plugin\Internal;
+
+return new MethodSearcherPlugin();